pyproject.toml
[project]
name = "taxlens"
version = "0.1.0"
description = "Global Accommodation Tax Intelligence Platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.9",
    "anthropic>=0.40.0",
    "apscheduler>=3.10.0",
    "croniter>=1.4.0",
    "beautifulsoup4>=4.12.0",
    "passlib[bcrypt]>=1.7.0",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"scripts/seed_data.py" = ["E501"]
"app/core/rule_engine.py" = ["E501"]
"app/api/tax_rules.py" = ["E501"]
"app/services/monitoring_service.py" = ["E501"]
"scripts/seed_etg_markets.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short"
markers = [
    "live: tests that call external APIs (deselect with -m 'not live')",
    "integration: tests requiring a running database",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
